BUGFIX handling negative inventory count when placing order by wallet

+++ weavewhisper/src/main/java/com/weavewhisper/services/impl/CartServiceImpl.java
@@ -264,6 +264,7 @@ public class CartServiceImpl implements CartService {
 	}
 
 	@Override
+	@Transactional
 	public ApiResponse handlePlaceOrderByWallet(PlaceOrderRequestDto placeOrderRequestDto) {
 
 		Customer customer = customerDao.findById(placeOrderRequestDto.getCustomerId())
@@ -275,44 +276,51 @@ public class CartServiceImpl implements CartService {
 
 		double price = 0;
 
-		for (int i = 0; i < cartList.size(); i++) {
-			Cart c = cartList.get(i);
-			if (c.getProductRef().getManufacturer() == null) {
-				throw new IllegalCartItemException("Some product from your cart doesnt exists anymore.");
-			} else if (c.getProductRef().getInventoryCount() == 0) {
-				throw new IllegalCartItemException("Some product from your cart is already sold out.");
-			}
-			price += c.getProductRef().getSellingPrice();
-		}
-
-		if (customer.getBalance() >= price) {
-			String phoneNumber = placeOrderRequestDto.getPhoneNumber();
-			String address = placeOrderRequestDto.getAddress();
-
+		try {
 			for (int i = 0; i < cartList.size(); i++) {
 				Cart c = cartList.get(i);
-				OrderHistory orderHistory = new OrderHistory();
-				orderHistory.setCustomerRef(customer);
-				orderHistory.setProductRef(c.getProductRef());
-				orderHistory.setManufacturer(c.getProductRef().getManufacturer());
-				orderHistory.setColor(c.getColor());
-				orderHistory.setSize(c.getSize());
-				orderHistory.setSoldAtPrice(c.getProductRef().getSellingPrice());
-				orderHistory.setPaymentType(PaymentType.WALLET);
-				orderHistory.setAddress(address);
-				orderHistory.setPhoneNumber(phoneNumber);
-				orderHistory.setReceipt(reciept);
-				System.out.println(orderHistory);
-				orderHistoryDao.save(orderHistory);
-				orderHistory.getProductRef().setInventoryCount(orderHistory.getProductRef().getInventoryCount() - 1);
-				cartDao.delete(c);
-
+				if (c.getProductRef().getManufacturer() == null) {
+					throw new IllegalCartItemException("Some product from your cart doesnt exists anymore.");
+				} else if (c.getProductRef().getInventoryCount() == 0) {
+					throw new IllegalCartItemException("Some product from your cart is already sold out.");
+				}
+				price += c.getProductRef().getSellingPrice();
 			}
-			customer.setBalance(customer.getBalance() - price);
-			return new ApiResponse(true, "Products successfully purchased..");
-		} else {
-			throw new LowBalanceException("Insufficient wallet balance.");
-		}
 
+			if (customer.getBalance() >= price) {
+				String phoneNumber = placeOrderRequestDto.getPhoneNumber();
+				String address = placeOrderRequestDto.getAddress();
+
+				for (int i = 0; i < cartList.size(); i++) {
+					Cart c = cartList.get(i);
+					OrderHistory orderHistory = new OrderHistory();
+					orderHistory.setCustomerRef(customer);
+					orderHistory.setProductRef(c.getProductRef());
+					orderHistory.setManufacturer(c.getProductRef().getManufacturer());
+					orderHistory.setColor(c.getColor());
+					orderHistory.setSize(c.getSize());
+					orderHistory.setSoldAtPrice(c.getProductRef().getSellingPrice());
+					orderHistory.setPaymentType(PaymentType.WALLET);
+					orderHistory.setAddress(address);
+					orderHistory.setPhoneNumber(phoneNumber);
+					orderHistory.setReceipt(reciept);
+					System.out.println(orderHistory);
+					orderHistoryDao.save(orderHistory);
+					if(orderHistory.getProductRef().getInventoryCount() == 0) {
+						throw new IllegalCartItemException("You are trying to buy an item which is no longer avaiable");
+					}
+					orderHistory.getProductRef().setInventoryCount(orderHistory.getProductRef().getInventoryCount() - 1);
+					cartDao.delete(c);
+
+				}
+				customer.setBalance(customer.getBalance() - price);
+				return new ApiResponse(true, "Products successfully purchased..");
+			} else {
+				throw new LowBalanceException("Insufficient wallet balance.");
+			}
+		} catch(IllegalCartItemException e) {
+			throw new PlaceOrderFailureException(
+					"Something went wrong. Any amount deducted from your account will get refunded withing 4-7 business days.");
+		}
 	}
 }
